Update Event Start
