+++ Components/EventCard/EventCard.js
@@ -1,5 +1,4 @@
 import {
-  Typography,
   Button,
   Card,
   CardContent,
@@ -7,28 +6,40 @@ import {
   CardMedia,
   CardHeader,
 } from '@mui/material';
+import { useRouter } from 'next/router';
 
-const EventCard = ({ id, openDialog }) => {
-  // Design will be changed, This is a basic one
+const EventCard = ({ id, openDialog, image, event_name, event_id }) => {
+  const router = useRouter();
+
+  const openEvent = () => {
+    router.push(`/eventList/${event_id}`);
+  };
+
   return (
     <Card variant="outlined">
       <CardContent>
-        <CardHeader>
-          <Typography sx={{ fontSize: 20 }}>Event Name</Typography>
-        </CardHeader>
+        <CardHeader
+          titleTypographyProps={{ fontSize: `1rem`, textAlign: `center` }}
+          title={event_name}
+        ></CardHeader>
         <CardMedia
           component="img"
           height="194"
-          image="https://smaller-pictures.appspot.com/images/dreamstime_xxl_65780868_small.jpg"
+          image={`${process.env.NEXT_PUBLIC_BACKEND_API}${image}`}
           alt="Event Photo"
         />
       </CardContent>
       <CardActions sx={{ justifyContent: 'center' }}>
-        <Button variant="contained" size="small" onClick={openDialog}>
+        <Button variant="contained" size="small" onClick={openDialog} id={id}>
           Edit Information
         </Button>
-        <Button variant="outlined" size="small">
-          Delete Event
+        <Button
+          variant="contained"
+          name={`${event_id}`}
+          size="small"
+          onClick={openEvent}
+        >
+          View Event
         </Button>
       </CardActions>
     </Card>
